Added test to demonstrate that maybe skip ngrams should have a skip size limit.
Update issue 89
Status: Started

Thanks for the new algorithm.
I noticed the new algorithm doesn't limit the size of the skip, other than the size of the Sentence.  This means that skip bigrams from a sentence with 40 tokens will include a bigram with token 1 and token 40.

I see two problems with this approach: 
1. It will generate an enormous number of skip ngrams.
2. This isn't how I understand skip ngrams to be implemented in literature such as Guthrie et al (2006), who impose an upper limit on the number of tokens skipped.

I am committing a unit test to illustrate this issue.