Adding a test for LucenePOSNGramFeatureExtractor.
Update issue 137
The class hierarchy of the LuceneNgram Feature Extractors was inconsistent. I changed it to be consistent with LucenePOSNGram FeatureExtractors etc. 
LucenePOSNGramFeatureExtractor has been working fine as shown in the corresponding test, but you need to set different parameters to configure it (as also shown in the test). I guess, this was causing the unexpected results here.