For each CV run, also create an aggregate OutcomeIdReport.
Update issue 143
I'm not sure whether we should really open the "mid" report level to any kind of report users could add. This is a quite complex matter for users who are not familiar with the task and report structure (especially because adding a report at mid level might require adding another report at low etc.). I therefore added OutcomeIdReport (the report which caused this issue) to BatchTrainTestReport (we are currently using this report as "mid-level" report).