Fixes issue 163.
SimpleDkproTCReader was not that simple at all, as it inherited from the CasResourceCollectionReader_ImplBase.
The resource-based layout does not make sense in this case, as only one file is read anyway.
This also caused the problem with the language parameter, as this was inherited from the base class, but proper cas initialization was not performed.

I have simplified that a bit and it should work now.