Fixes issue 201.
Introduced feature store filters than are applied before the feature store is written.

Added UniformClassDistributionFilter (similar to SpreadSubsampleFilter in Weka) that is applied to the training data in order to train a model on balanced class distribution (but might discard instances for that).

Possible other filters might be oversampling "filter" that oversample the small classes (not implemented yet).
More filters can be added at any time.

Also added a demo that shows how to use filters.